+++ rut/src/test/java/io/danielnorberg/rut/RadixTrieTest.java
@@ -1,4 +1,4 @@
-package dano;
+package io.danielnorberg.rut;
 
 import org.junit.Test;
